Quick May Head Changes
After trying to get the heads, I noticed that both Warden head and the Bee heads are very hard to get.
On one side, wardens take a lot to kill.
One the other side, theres 7 bee heads to get with only a 6% chance, so it is very hard to get all of them.
Changed both of these to 10% to make it a little more balanced, but not too easy.